+++ src/main/java/com/ksolorzano/KinScription/web/controller/AdminInscripcionController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Controller
@@ -40,7 +41,7 @@ public class AdminInscripcionController {
      */
     @PostMapping("/calificar")
     public String calificarExamen(@RequestParam("participanteId") int participanteId,
-                                  @RequestParam("nota") double nota) {
+                                  @RequestParam("nota") BigDecimal nota) {
         try {
             participanteService.calificarExamen(participanteId, nota);
         } catch (Exception e) {
